requirements.txt
Flask
numpy
pillow
h5py
tensorflow